src/services/market/requirements.txt
flask == 3.0.3
psycopg2-binary == 2.9.10
bcrypt == 4.2.0